+++ CHANGELOG.rst
@@ -3,7 +3,13 @@ Change Log
 ==========
 
 ``master`` (unreleased)
--------------------------
+-----------------------
+
+* WIP
+
+
+``0.0.11`` (2017-08-23)
+-----------------------
 
 * New tab shortcut changed to ``Cmd-T`` (reported by Ole)
 * Open quickly shortcut synced with Xcode to ``Cmd-Shift-O`` (reported by Ole)